requirements.txt
fastapi
uvicorn
python-dotenv
pydantic[email]==1.10.13
sqlalchemy>=1.4
asyncpg
bcrypt
PyJWT
pytest
httpx
fastapi-jwt-auth
sqlalchemy
alembic
psycopg2-binary
email-validator

boto3
python-multipart
